Move maps functions to maps.js
